+++ Trabalho2909 - Continuacao_parte3/atividade_2909/aula1509/src/main/java/pweb/aula2909/model/entity/ItemVenda.java
@@ -24,7 +24,7 @@ public class ItemVenda {
     @Column(nullable = false)
     @NotNull(message = "Quantidade é obrigatória")
     @DecimalMin(value = "1", message = "Quantidade deve ser maior que zero")
-    private Integer quantidade;
+    private Double quantidade;
 
     public Long getId() {
         return id;
@@ -50,11 +50,11 @@ public class ItemVenda {
         this.venda = venda;
     }
 
-    public Integer getQuantidade() {
+    public Double getQuantidade() {
         return quantidade;
     }
 
-    public void setQuantidade(Integer quantidade) {
+    public void setQuantidade(Double quantidade) {
         this.quantidade = quantidade;
     }
